requirements.txt
# 基础工具
duckduckgo_search>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=1.0.0

# LLM & RAG 核心组件
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-chroma>=0.1.0    # <--- 这次报错缺失的包
chromadb>=0.4.0
unstructured>=0.11.0

# 文档处理
pypdf
python-docx
streamlit